+++ src/main/java/com/devcard/devcard/auth/controller/rest/OauthController.java
@@ -1,4 +1,4 @@
-package com.devcard.devcard.auth.controller;
+package com.devcard.devcard.auth.controller.rest;
 
 import com.devcard.devcard.auth.entity.Member;
 import com.devcard.devcard.auth.model.OauthMemberDetails;
@@ -6,10 +6,8 @@ import com.devcard.devcard.auth.repository.MemberRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.security.core.Authentication;
 
 import java.util.HashMap;
 import java.util.Map;
